+++ packages/gatsby/src/bootstrap/requires-writer.ts
@@ -163,8 +163,6 @@ const getMatchPaths = (
 const createHash = (
   matchPaths: Array<IGatsbyPageMatchPath>,
   components: Array<IGatsbyPageComponent>,
-  cleanedClientVisitedPageComponents: Array<IGatsbyPageComponent>,
-  notVisitedPageComponents: Array<IGatsbyPageComponent>,
   cleanedSSRVisitedPageComponents: Array<IGatsbyPageComponent>
 ): string =>
   crypto
@@ -173,8 +171,6 @@ const createHash = (
       JSON.stringify({
         matchPaths,
         components,
-        cleanedClientVisitedPageComponents,
-        notVisitedPageComponents,
         cleanedSSRVisitedPageComponents,
       })
     )
@@ -199,37 +195,9 @@ export const writeAll = async (state: IGatsbyState): Promise<boolean> => {
     )
   }
 
-  let cleanedClientVisitedPageComponents: Array<IGatsbyPageComponent> = []
-  let notVisitedPageComponents: Array<IGatsbyPageComponent> = []
-
-  if (process.env.GATSBY_EXPERIMENTAL_LAZY_DEVJS) {
-    const clientVisitedPageComponents = [
-      ...(state.visitedPages.get(`client`)?.values() || []),
-    ]
-
-    // Remove any page components that no longer exist.
-    cleanedClientVisitedPageComponents = components.filter(component =>
-      clientVisitedPageComponents.some(
-        pageComponentChunkName =>
-          pageComponentChunkName === component.componentChunkName
-      )
-    )
-
-    // Get list of page components that the user has _not_ visited.
-    notVisitedPageComponents = components.filter(
-      component =>
-        // Filter out page components the user has visited.
-        !cleanedClientVisitedPageComponents.some(
-          c => c.componentChunkName === component.componentChunkName
-        )
-    )
-  }
-
   const newHash = createHash(
     matchPaths,
     components,
-    cleanedClientVisitedPageComponents,
-    notVisitedPageComponents,
     cleanedSSRVisitedPageComponents
   )
 
@@ -282,35 +250,6 @@ const preferDefault = m => (m && m.default) || m
     )
     .join(`,\n`)}
 }\n\n`
-
-  if (process.env.GATSBY_EXPERIMENTAL_LAZY_DEVJS) {
-    // Create file with sync requires of visited page components files.
-    let lazyClientSyncRequires = `${hotImport}
-  // prefer default export if available
-  const preferDefault = m => (m && m.default) || m
-  \n\n`
-    lazyClientSyncRequires += `exports.lazyComponents = {\n${cleanedClientVisitedPageComponents
-      .map(
-        (c: IGatsbyPageComponent): string =>
-          `  "${
-            c.componentChunkName
-          }": ${hotMethod}(preferDefault(require("${joinPath(c.component)}")))`
-      )
-      .join(`,\n`)}
-  }\n\n`
-
-    // Add list of not visited components.
-    lazyClientSyncRequires += `exports.notVisitedPageComponents = {\n${notVisitedPageComponents
-      .map(
-        (c: IGatsbyPageComponent): string => `  "${c.componentChunkName}": true`
-      )
-      .join(`,\n`)}
-  }\n\n`
-
-    writeModule(`$virtual/lazy-client-sync-requires`, lazyClientSyncRequires)
-  } else {
-    writeModule(`$virtual/lazy-client-sync-requires`, ``)
-  }
 
   // Create file with async requires of components/json files.
   let asyncRequires = `// prefer default export if available
@@ -382,33 +321,29 @@ const debouncedWriteAll = _.debounce(
   }
 )
 
-if (process.env.GATSBY_EXPERIMENTAL_LAZY_DEVJS) {
-  /**
-   * Start listening to CREATE_CLIENT_VISITED_PAGE events so we can rewrite
-   * files as required
-   */
-  emitter.on(`CREATE_CLIENT_VISITED_PAGE`, (): void => {
-    reporter.pendingActivity({ id: `requires-writer` })
-    debouncedWriteAll()
-  })
-}
-
-if (process.env.GATSBY_EXPERIMENTAL_DEV_SSR) {
-  /**
-   * Start listening to CREATE_SERVER_VISITED_PAGE events so we can rewrite
-   * files as required
-   */
-  emitter.on(`CREATE_SERVER_VISITED_PAGE`, (): void => {
-    reporter.pendingActivity({ id: `requires-writer` })
-    debouncedWriteAll()
-  })
-}
-
 /**
  * Start listening to CREATE/DELETE_PAGE events so we can rewrite
  * files as required
  */
+let listenerStarted = false
 export const startListener = (): void => {
+  // Only start the listener once.
+  if (listenerStarted) {
+    return
+  }
+  listenerStarted = true
+
+  if (process.env.GATSBY_EXPERIMENTAL_DEV_SSR) {
+    /**
+     * Start listening to CREATE_SERVER_VISITED_PAGE events so we can rewrite
+     * files as required
+     */
+    emitter.on(`CREATE_SERVER_VISITED_PAGE`, (): void => {
+      reporter.pendingActivity({ id: `requires-writer` })
+      debouncedWriteAll()
+    })
+  }
+
   emitter.on(`CREATE_PAGE`, (): void => {
     reporter.pendingActivity({ id: `requires-writer` })
     debouncedWriteAll()