+++ packages/gatsby/src/redux/reducers/visited-page.ts
@@ -1,11 +1,10 @@
 import {
   IGatsbyState,
   IDeleteCacheAction,
-  ICreateClientVisitedPage,
   ICreateServerVisitedPage,
 } from "../types"
 
-type StateMap = Map<"client" | "server", Set<string>>
+type StateMap = Map<"server", Set<string>>
 
 // The develop server always wants these page components.
 const createDefault = (): StateMap => {
@@ -14,33 +13,18 @@ const createDefault = (): StateMap => {
   defaults.add(`component---src-pages-404-js`)
   defaults.add(`component---src-pages-index-js`)
 
-  const state: StateMap = new Map([
-    [`client`, new Set(defaults)],
-    [`server`, new Set(defaults)],
-  ])
+  const state: StateMap = new Map([[`server`, new Set(defaults)]])
 
   return state
 }
 
 export const visitedPagesReducer = (
   state: IGatsbyState["visitedPages"] = createDefault(),
-  action:
-    | IDeleteCacheAction
-    | ICreateClientVisitedPage
-    | ICreateServerVisitedPage
+  action: IDeleteCacheAction | ICreateServerVisitedPage
 ): IGatsbyState["visitedPages"] => {
   switch (action.type) {
     case `DELETE_CACHE`:
       return createDefault()
-
-    case `CREATE_CLIENT_VISITED_PAGE`: {
-      const client = state.get(`client`)
-      if (client) {
-        client.add(action.payload.componentChunkName)
-      }
-
-      return state
-    }
 
     case `CREATE_SERVER_VISITED_PAGE`: {
       const server = state.get(`server`)